Split loss tests out of functional tests
