+++ scico/test/test_functional.py
@@ -10,55 +10,17 @@ import warnings
 import jax
 
 import pytest
+from prox import prox_test
 
 import scico.numpy as snp
-from scico import denoiser, functional, linop, loss
+from scico import denoiser, functional
 from scico.blockarray import BlockArray
 from scico.random import randn
-from scico.solver import minimize
 
 NO_BLOCK_ARRAY = [functional.L21Norm, functional.NuclearNorm]
 NO_COMPLEX = [
     functional.NonNegativeIndicator,
 ]
-
-
-def prox_func(x, v, f, alpha):
-    """Evaluate functional of which the proximal operator is the argmin."""
-    return 0.5 * snp.sum(snp.abs(x.reshape(v.shape) - v) ** 2) + alpha * snp.array(
-        f(x.reshape(v.shape)), dtype=snp.float64
-    )
-
-
-def prox_solve(v, v0, f, alpha):
-    """Evaluate the alpha-scaled proximal operator of f at v, using v0 as an
-    initial point for the optimization."""
-    fnc = lambda x: prox_func(x, v, f, alpha)
-    fmn = minimize(
-        fnc,
-        v0,
-        method="Nelder-Mead",
-        options={"maxiter": 1000, "xatol": 1e-9, "fatol": 1e-9},
-    )
-
-    return fmn.x.reshape(v.shape), fmn.fun
-
-
-def prox_test(v, nrm, prx, alpha, x0=None):
-    """Test the alpha-scaled proximal operator function prx of norm functional nrm
-    at point v."""
-    # Evaluate the proximal operator at v
-    px = snp.array(prx(v, alpha, v0=x0))
-    # Proximal operator functional value (i.e. Moreau envelope) at v
-    pf = prox_func(px, v, nrm, alpha)
-    # Brute-force solve of the proximal operator at v
-    mx, mf = prox_solve(v, px, nrm, alpha)
-
-    # Compare prox functional value with brute-force solution
-    if pf < mf:
-        return  # prox gave a lower cost than brute force, so it passes
-
-    np.testing.assert_allclose(pf, mf, rtol=1e-6)
 
 
 class ProxTestObj:
@@ -322,105 +284,6 @@ def test_scalar_pmap():
     non_pmap = np.array([foo(c) for c in c_list])
     pmapped = jax.pmap(foo)(c_list)
     np.testing.assert_allclose(non_pmap, pmapped)
-
-
-class TestLoss:
-    def setup_method(self):
-        n = 4
-        dtype = np.float64
-        A, key = randn((n, n), key=None, dtype=dtype, seed=1234)
-        D, key = randn((n,), key=key, dtype=dtype)
-        W, key = randn((n,), key=key, dtype=dtype)
-        W = 0.1 * W + 1.0
-        self.Ao = linop.MatrixOperator(A)
-        self.Ao_abs = linop.MatrixOperator(snp.abs(A))
-        self.Do = linop.Diagonal(D)
-        self.W = linop.Diagonal(W)
-        self.y, key = randn((n,), key=key, dtype=dtype)
-        self.v, key = randn((n,), key=key, dtype=dtype)  # point for prox eval
-        scalar, key = randn((1,), key=key, dtype=dtype)
-        self.scalar = scalar.copy().ravel()[0]
-
-    def test_squared_l2(self):
-        L = loss.SquaredL2Loss(y=self.y, A=self.Ao)
-        assert L.has_eval == True
-        assert L.has_prox == True
-
-        # test eval
-        np.testing.assert_allclose(L(self.v), 0.5 * ((self.Ao @ self.v - self.y) ** 2).sum())
-
-        cL = self.scalar * L
-        assert L.scale == 0.5  # hasn't changed
-        assert cL.scale == self.scalar * L.scale
-        assert cL(self.v) == self.scalar * L(self.v)
-
-        # SquaredL2 with Diagonal linop has a prox
-        L_d = loss.SquaredL2Loss(y=self.y, A=self.Do)
-
-        # test eval
-        np.testing.assert_allclose(L_d(self.v), 0.5 * ((self.Do @ self.v - self.y) ** 2).sum())
-
-        assert L_d.has_eval == True
-        assert L_d.has_prox == True
-
-        cL = self.scalar * L_d
-        assert L_d.scale == 0.5  # hasn't changed
-        assert cL.scale == self.scalar * L_d.scale
-        assert cL(self.v) == self.scalar * L_d(self.v)
-
-        pf = prox_test(self.v, L_d, L_d.prox, 0.75)
-
-        pf = prox_test(self.v, L, L.prox, 0.75)
-
-    def test_weighted_squared_l2(self):
-        L = loss.WeightedSquaredL2Loss(y=self.y, A=self.Ao, W=self.W)
-        assert L.has_eval == True
-        assert L.has_prox == True
-
-        # test eval
-        np.testing.assert_allclose(
-            L(self.v), 0.5 * (self.W @ (self.Ao @ self.v - self.y) ** 2).sum()
-        )
-
-        cL = self.scalar * L
-        assert L.scale == 0.5  # hasn't changed
-        assert cL.scale == self.scalar * L.scale
-        assert cL(self.v) == self.scalar * L(self.v)
-
-        # SquaredL2 with Diagonal linop has a prox
-        L_d = loss.WeightedSquaredL2Loss(y=self.y, A=self.Do, W=self.W)
-
-        assert L_d.has_eval == True
-        assert L_d.has_prox == True
-
-        # test eval
-        np.testing.assert_allclose(
-            L_d(self.v), 0.5 * (self.W @ (self.Do @ self.v - self.y) ** 2).sum()
-        )
-
-        cL = self.scalar * L_d
-        assert L_d.scale == 0.5  # hasn't changed
-        assert cL.scale == self.scalar * L_d.scale
-        assert cL(self.v) == self.scalar * L_d(self.v)
-
-        pf = prox_test(self.v, L_d, L_d.prox, 0.75)
-
-        pf = prox_test(self.v, L, L.prox, 0.75)
-
-    def test_poisson(self):
-        L = loss.PoissonLoss(y=self.y, A=self.Ao_abs)
-        assert L.has_eval == True
-        assert L.has_prox == False
-
-        # test eval
-        v = snp.abs(self.v)
-        Av = self.Ao_abs @ v
-        np.testing.assert_allclose(L(v), 0.5 * snp.sum(Av - self.y * snp.log(Av) + L.const))
-
-        cL = self.scalar * L
-        assert L.scale == 0.5  # hasn't changed
-        assert cL.scale == self.scalar * L.scale
-        assert cL(v) == self.scalar * L(v)
 
 
 class TestBM3D:
